Update App.js
I did a few things here, I created the function createFingerprint() (which is called in the wrong place, don't know how to put in the correct place)

Here is the order of operations

createFingerprint() - creates a unique fingerprint hash
register() - sets up the socket io connection to the server
stompClinetReady() - socket io connection is ready, lets call initwallet()
initWallet() - this sends a signal to the server that everything is ready to go, lets send this wallet money if balance == 0
walletUpdate() - this is a special function called by the socket io manager. This will get called ANYTIME there is a change to the wallet ( less money, more money). The states need to updated here